feat: 삭제/수정 불가능한 테마에 대한 커스텀 예외 추가 및 적용

+++ src/main/java/nextstep/web/theme/service/ThemeService.java
@@ -1,6 +1,10 @@
 package nextstep.web.theme.service;
 
+import nextstep.domain.Reservation;
 import nextstep.domain.Theme;
+import nextstep.web.common.exception.BusinessException;
+import nextstep.web.common.exception.CommonErrorCode;
+import nextstep.web.reservation.repository.ReservationDao;
 import nextstep.web.theme.dto.*;
 import nextstep.web.common.repository.RoomEscapeRepository;
 import nextstep.web.theme.repository.ThemeDao;
@@ -12,10 +16,13 @@ import org.springframework.stereotype.Service;
 public class ThemeService {
 
     private final RoomEscapeRepository<Theme> themeRepository;
+    private final RoomEscapeRepository<Reservation> reservationRepository;
 
     @Autowired
-    public ThemeService(@Qualifier("themeDao") RoomEscapeRepository<Theme> themeRepository) {
+    public ThemeService(@Qualifier("themeDao") RoomEscapeRepository<Theme> themeRepository,
+                        @Qualifier("reservationDao") RoomEscapeRepository<Reservation> reservationRepository) {
         this.themeRepository = themeRepository;
+        this.reservationRepository = reservationRepository;
     }
 
     public CreateThemeResponseDto createTheme(CreateThemeRequestDto requestDto) {
@@ -35,10 +42,22 @@ public class ThemeService {
     }
 
     public void deleteTheme(Long id) {
+        if (isReserved(id)) {
+            throw new BusinessException(CommonErrorCode.RESERVED_THEME_ERROR);
+        }
         themeRepository.deleteById(id);
     }
 
     public void updateTheme(CreateThemeRequestDto requestDto, Long id) {
-        ((ThemeDao) themeRepository).updateById(Theme.of(requestDto, id));
+        if (isReserved(id)) {
+            throw new BusinessException(CommonErrorCode.RESERVED_THEME_ERROR);
+        }
+        if (((ThemeDao) themeRepository).updateById(Theme.of(requestDto, id)) == 0) {
+            throw new BusinessException(CommonErrorCode.RESOURCE_NOT_FOUND);
+        }
+    }
+
+    private boolean isReserved(Long themeId) {
+        return ((ReservationDao) reservationRepository).findByThemeId(themeId).size() == 0;
     }
 }
